Finished implementing Freebie Tracker feature
The connectors check{in,out}_field methods now return a dict which includes
a status message in 'message' and the signup object in 'signup'. This is now
used in the checkin/checkin_update_data and /mutate to generate the appropriate
response to the user. The connectors can now chose their own message.

The javascript on the checkin page was updated to also handle the freebies_taken
property and display it in the Status column.

When creating a new freebie tracker, the entered value for maximum freebies
allowed is checked to be greater or equal to 1.